# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: API de Gestión de Transacciones Bancarias
  description: API para registrar y gestionar transacciones bancarias, incluyendo depósitos, retiros, transferencias e historial.
  version: 1.0.0
servers:
  - url: http://localhost:8090
    description: Servidor local
paths:
  /transacciones/deposito:
    post:
      summary: Registrar un depósito
      description: Registrar una transacción de depósito entre dos cuentas bancarias.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositoRequest'
            examples:
              deposito:
                value:
                  sourceAccountId: "12345"
                  destinationAccountId: "67890"
                  amount: 1000
      responses:
        '201':
          description: Depósito registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaTransaccion'
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaTransaccion'
  /transacciones/retiro:
    post:
      summary: Registrar un retiro
      description: Registrar una transacción de retiro en una cuenta bancaria.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetiroRequest'
            examples:
              retiro:
                value:
                  accountId: "12345"
                  amount: 500
      responses:
        '201':
          description: Retiro registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaTransaccion'
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaTransaccion'
  /transacciones/transferencia:
    post:
      summary: Registrar una transferencia
      description: Registrar una transferencia entre dos cuentas bancarias.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferenciaRequest'
            examples:
              transferencia:
                value:
                  sourceAccountId: "12345"
                  destinationAccountId: "67890"
                  amount: 1500
      responses:
        '201':
          description: Transferencia registrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaTransaccion'
  /transacciones/historial:
    get:
      summary: "Consultar historial de transacciones"
      description: "Devuelve el historial de todas las transacciones registradas, con opción de filtrado."
      operationId: "transaccionesHistorialGet"
      parameters:
        - name: "id"
          in: "query"
          description: "ID de la cuenta asociada"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: "Historial de transacciones"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/RespuestaTransaccion"
        '400':
          description: "Error en la solicitud"
        '500':
          description: "Error interno del servidor"
components:
  schemas:
    DepositoRequest:
      type: object
      properties:
        sourceAccountId:
          type: string
          description: ID de la cuenta de origen del depósito.
        destinationAccountId:
          type: string
          description: ID de la cuenta de destino del depósito.
        amount:
          type: number
          format: double
          description: Monto del depósito.
      required:
        - sourceAccountId
        - destinationAccountId
        - amount
    RetiroRequest:
      type: object
      properties:
        accountId:
          type: string
          description: ID de la cuenta asociada al retiro.
        amount:
          type: number
          format: double
          description: Monto del retiro.
      required:
        - accountId
        - amount
    TransferenciaRequest:
      type: object
      properties:
        sourceAccountId:
          type: string
          description: ID de la cuenta de origen.
        destinationAccountId:
          type: string
          description: ID de la cuenta de destino.
        amount:
          type: number
          format: double
          description: Monto de la transferencia.
      required:
        - sourceAccountId
        - destinationAccountId
        - amount
    RespuestaTransaccion:
      type: object
      properties:
        transactionId:
          type: string
          description: ID único de la transacción.
        transactionType:
          type: string
          description: Tipo de transacción (deposito, retiro, transferencia).
        amount:
          type: number
          format: double
          description: Monto de la transacción.
        date:
          type: string
          format: date-time
          description: Fecha de la transacción.
        accountId:
          type: string
          description: ID de la cuenta asociada.
        destinationAccountId:
          type: string
          nullable: true
          description: ID de la cuenta de destino (para transferencias y depósitos).
      required:
        - transactionId
        - transactionType
        - amount
        - date